+++ templates/home.htm
@@ -1,0 +1,9 @@
+<!doctype html>
+<html>
+   <body>
+       <h1>Total Cases & Growth Rate</h1>
+       <img src="/growth.png" alt="Cases Growth Rate">
+       <h1>State wise Death Rate</h1>
+       <img src="/death.png" alt="State wise Death Rate">
+   </body>
+</html>
